refactor : [FileUtil] Builder pattern 적용, 사용자 예외 처리 설정. header 설정 미적용.

+++ src/main/java/com/scyllacore/dumpWeb/commonModule/db/mapper/file/FileMapper.java
@@ -4,8 +4,8 @@ import com.scyllacore.dumpWeb.commonModule.db.dto.manage.FileDTO;
 public interface FileMapper {
     int insertFileInfoByGroupReportId(FileDTO fileDTO);
 
-    FileDTO findFileInfoByFileId(int fileId);
+    FileDTO findFileInfoByFileId(Long fileId);
 
-    void deleteFile(long fileId);
+    void deleteFile(Long fileId);
 
 }
